Update kemono-dl.py
added  --min-filesize, --max-filesize, --skip-content, --skip-embeds,  --favorite-users, --favorite-posts
fixed bug with printing date out of range
also minor tweaks

+++ kemono-dl.py
@@ -1,6 +1,4 @@
 import requests
-from bs4 import BeautifulSoup
-from http.cookiejar import MozillaCookieJar
 import os
 import re
 import argparse
@@ -9,8 +7,10 @@ import time
 import datetime
 import json
 from PIL import Image
+from bs4 import BeautifulSoup
+from http.cookiejar import MozillaCookieJar
 
-version = '2021.10.11.1'
+version = '2021.10.13'
 
 ap = argparse.ArgumentParser()
 ap.add_argument("--version", action='store_true', help="Displays the current version then exits")
@@ -21,13 +21,17 @@ ap.add_argument("-o", "--output", help="Set path to download posts")
 ap.add_argument("-a", "--archive", help="Downloads only posts that are not in provided archive file")
 ap.add_argument("-i", "--ignore-errors", action='store_true', help="Continue to download post(s) when an error occurs")
 ap.add_argument("-s", "--simulate", action='store_true', help="Print post(s) info and does not download")
-ap.add_argument("--date", help="Only download posts from this date")
-ap.add_argument("--datebefore", help="Only download posts from this date and before")
-ap.add_argument("--dateafter", help="Only download posts from this date and after")
-ap.add_argument("--force-inline", action='store_true', help="Force download all external inline images found in post content")
-# ap.add_argument("--min-filesize", help="Do not download files smaller than this")
-# ap.add_argument("--max-filesize", help="Do not download files larger than this")
-# ap.add_argument("-up", "--update", action='store_true', help="Redownloads any post that has been updated (ignores --archive)")
+ap.add_argument("--date", help="Only download posts from this date. (Format: YYYYMMDD)")
+ap.add_argument("--datebefore", help="Only download posts from this date and before. (Format: YYYYMMDD)")
+ap.add_argument("--dateafter", help="Only download posts from this date and after. (Format: YYYYMMDD)")
+ap.add_argument("--force-inline", action='store_true', help="Force download all external inline images found in post content. (experimental)")
+ap.add_argument("--min-filesize", help="Do not download files smaller than this. (Format: 1GB, 1MB, 1KB, 1B)")
+ap.add_argument("--max-filesize", help="Do not download files larger than this. (Format: 1GB, 1MB, 1KB, 1B)")
+ap.add_argument("--skip-content", action='store_true', help="Skips creating content.html")
+ap.add_argument("--skip-embeds", action='store_true', help="Skips creating external_links.txt")
+ap.add_argument("--favorite-users", action='store_true', help="Downloads all users saved in your favorites. (Requires --cookies)")
+ap.add_argument("--favorite-posts", action='store_true', help="Downloads all posts saved in your favorites. (Requires --cookies)")
+# ap.add_argument("-up", "--update", action='store_true', help="Redownloads any post that has been updated (ignores --archive)") # might need to create a log file of kemono.party last edit time
 args = vars(ap.parse_args())
 
 if args['version']: print(version), quit()
@@ -56,41 +60,57 @@ if args['archive']:
     archive_flag = True
     archive_file = args['archive']
 
-def validate_date(date:str):
-    try: 
-        datetime.datetime.strptime(date, r'%Y%m%d')
-        return True
-    except: 
-        print("Error incorrect data format, should be YYYYMMDD"), quit()   
+def valid_date(date:str):
+    try: datetime.datetime.strptime(date, r'%Y%m%d')  
+    except: print("Error incorrect data format, should be YYYYMMDD"), quit()
+    return True   
+     
+if args['date']: valid_date(args['date'])
+if args['datebefore']: valid_date(args['datebefore'])
+if args['dateafter']: valid_date(args['dateafter'])
 
-check_date_flag = False      
-if args['date']: check_date_flag = validate_date(args['date'])
+def valid_size(size:str):
+    giga = re.search('([0-9]+)GB', size)
+    mega = re.search('([0-9]+)MB', size)
+    kilo = re.search('([0-9]+)KB', size)
+    byte = re.search('([0-9]+)B', size)
+    if giga: return int(giga.group(1)) * 10**9
+    elif mega: return int(mega.group(1)) * 10**6
+    elif kilo: return int(kilo.group(1)) * 10**2
+    elif byte: return int(byte.group(1))
+    else: print("Error incorrect size format, should be 1GB, 1MB, 1KB, 1B"), quit()
 
-if args['datebefore']: check_date_flag = validate_date(args['datebefore'])
-
-if args['dateafter']: check_date_flag = validate_date(args['dateafter'])
+if args['max_filesize']: args['max_filesize'] = valid_size(args['max_filesize'])
+if args['min_filesize']: args['min_filesize'] = valid_size(args['min_filesize'])
     
-def check_date(post_date:str, date:str, date_before:str, date_after:str):
-    if post_date == '00000000': return False
-    if date_before == None: date_before = '00000000'
-    if date_after == None: date_after = '99999999'
-    if date and (date_before or date_after):
-        return True if int(post_date) == int(date) or int(post_date) <= int(date_before) or int(post_date) >= int(date_after) else False
-    elif date:
-        return True if int(post_date) == int(date) else False
-    elif date_before or date_after:
-        return True if int(post_date) <= int(date_before) or int(post_date) >= int(date_after) else False
-                
+def check_date(date:str):
+    if not args['date'] and not args['datebefore'] and not args['dateafter']: return True
+    if date == '00000000': return False
+    if not args['datebefore']: args['datebefore'] = '0'
+    if not args['dateafter']: args['dateafter'] = float('inf')
+    if not args['date']: args['date'] = '0'
+    return True if int(date) == int(args['date']) or int(date) <= int(args['datebefore']) or int(date) >= float(args['dateafter']) else False
+    
+def check_size(size:int):
+    if not args['min_filesize'] and not args['max_filesize']: return True
+    if size == 0: return False
+    if not args['min_filesize']: args['min_filesize'] = '0'
+    if not args['max_filesize']: args['max_filesize'] = float('inf')
+    return True if size <= float(args['max_filesize']) and size >= int(args['min_filesize']) else False    
+           
 def download_file(file_name:str, url:str, file_path:str):
     try:
-        if not os.path.exists(file_path):
-            os.makedirs(file_path)
         file_name = re.sub('[\\/:\"*?<>|]+','',file_name) # remove illegal windows characters from file name
         print('Downloading: {}'.format(file_name))    
         with requests.get(url,stream=True,cookies=cookie_jar) as r:
             r.raise_for_status()
             downloaded = 0
-            total = int(r.headers.get('content-length', 0))
+            total = int(r.headers.get('content-length', '0'))
+            if not check_size(total):
+                print('File size out of range: {} bytes'.format(total))
+                return True
+            if not os.path.exists(file_path):
+                os.makedirs(file_path)
             with open(os.path.join(file_path, file_name), 'wb') as f:
                 start = time.time()
                 for chunk in r.iter_content(chunk_size=max(int(total/1000), 1024*1024)):                   
@@ -160,13 +180,7 @@ def print_post_data(post:dict):
     if post['file']:
         print('Files:\n\tFile name: {name}\n\tFile Path: https://kemono.party/data/{path}'.format(**post['file']))
     print('Content:\n{}'.format(BeautifulSoup(post['content'], 'html.parser').getText(separator="\n")))
-    print('-' * 50)
-    
-def print_channel_post_data(post:dict):
-    for key, value in post.items():
-        print(key, ' : ', value)
-    print('-' * 50)
-    return       
+    print('-' * 50)     
 
 def extract_post(post:dict, info:dict):
     """
@@ -190,24 +204,22 @@ def extract_post(post:dict, info:dict):
             ['subject']     # str
             ['url']         # str
     """
-    archived = []
-    if archive_flag:
-        if os.path.exists(archive_file):
-            with open(archive_file,'r') as f:
-                archived = f.read().splitlines() 
-               
     errors = 0
     info['post_id'] = post['id']
-         
+    
+    archived = []
+    if archive_flag and os.path.exists(archive_file):
+        with open(archive_file,'r') as f:
+            archived = f.read().splitlines() 
+               
     if not '{service} {user_id} {post_id}'.format(**info) in archived:
         
         try: date = datetime.datetime.strptime(post['published'], r'%a, %d %b %Y %H:%M:%S %Z').strftime(r'%Y%m%d')   
         except: date = '00000000'
         
-        if check_date_flag:
-            if not check_date(date, args['date'], args['datebefore'], args['dateafter']):
-                print('Date out of range: {} service: [{service}] user_id: [{user_id}] post_id: [{pos_id}]'.format(date, **info))
-                return        
+        if not check_date(date):
+            print('Date out of range: {} service: [{service}] user_id: [{user_id}] post_id: [{post_id}]'.format(date, **info))
+            return        
     
         if simulation_flag:
             print_post_data(post)
@@ -215,30 +227,27 @@ def extract_post(post:dict, info:dict):
     
         post_title = re.sub('[\\n\\t]+',' ', re.sub('[\\/:\"*?<>|]+','', post['title'] )).strip('.').strip() # removing illegal windows characters
         post_path = os.path.join(download_location, info['service'], '{username} [{user_id}]'.format(**info), '[{}] [{post_id}] {}'.format(date, post_title, **info))
-        if not os.path.exists(post_path):
-            os.makedirs(post_path)
-                
-        if not post['content'] == '':
+        
+        if post['content'] and not args['skip_content']:
+            if not os.path.exists(post_path):
+                os.makedirs(post_path)
             result = download_inline(post['content'], post_path, args['force_inline'])
             errors += result[1]
-            # write content to file 
             with open(os.path.join(post_path, 'content.html'),'wb') as File:
                 File.write(result[0].prettify().encode("utf-16"))
                 
-        # download post attachments
         for item in post['attachments']:
             errors += 1 if not download_file(item['name'], 'https://kemono.party/data{path}'.format(**item), os.path.join(post_path, 'attachments')) else 0
                 
-        # download post file
         if post['file']:
             errors += 1 if not download_file(post['file']['name'], 'https://kemono.party/data{path}'.format(**post['file']), post_path) else 0
                  
-        # save embedded links
-        if post['embed']:
+        if post['embed'] and not args['skip_embeds']:
+            if not os.path.exists(post_path):
+                os.makedirs(post_path)
             with open(os.path.join(post_path, 'external_links.txt'),'wb') as f:
                 f.write('{subject}\n{url}\n{description}'.format(**post['embed']).encode("utf-16"))
-                
-        # check total errors            
+                           
         if not errors:
             if archive_flag:
                 with open(archive_file,'a') as f:
@@ -249,6 +258,12 @@ def extract_post(post:dict, info:dict):
         return
     print("Already archived post. service: [{service}] user_id: [{user_id}] post_id: [{post_id}]".format(**info))
     return    
+
+def print_channel_post_data(post:dict):
+    for key, value in post.items():
+        print(key, ' : ', value)
+    print('-' * 50)
+    return  
 
 def extract_channel_post(post:dict, info:dict, channel:dict):
     """
@@ -303,6 +318,9 @@ def get_discord_chanels(info:dict):
     return json.loads(api_responce.text)
 
 def get_posts(info:dict):
+    if info['service'] == 'discord':
+        print('Saving Discords is still being developed')
+        return
     channels = get_discord_chanels(dict(info)) if info['service'] == 'discord' else [0]
     for channel in channels:
         chunk = 0
@@ -339,9 +357,9 @@ def get_pfp_banner(info:dict):
             try:
                 with Image.open(os.path.join(folder_path, file_name)) as image:
                     image.save(os.path.join(folder_path, '{}.{}'.format(file_name, image.format.lower())), format=image.format)
-                os.remove(os.path.join(folder_path, file_name)) 
-            except:
-                os.remove(os.path.join(folder_path, file_name)) # site might return garbage data if no icon or banner is found
+            except: pass
+            if os.path.exists(os.path.join(folder_path, file_name)):
+                os.remove(os.path.join(folder_path, file_name))
     return
 
 def get_username(service:str, user_id:str):
@@ -373,7 +391,59 @@ def extract_link(link:str):
         return True
     return False    
 
+def get_favorite_users():
+    try:
+        api_favorites = 'https://kemono.party/api/favorites?type=artist'
+        api_responce = requests.get(api_favorites, cookies=cookie_jar)
+        api_responce.raise_for_status()
+        data = json.loads(api_responce.text)
+        if not data:
+            print('You have no favorite users.')
+            return
+        for favorite_user in data:
+            """
+            favorite_user       # dict
+                ['faved_seq']   # int
+                ['id']          # str
+                ['indexed']     # str, datetime object
+                ['name']        # str
+                ['service']     # str
+                ['updated']     # str, datetime object
+            """
+            extract_link('https://kemono.party/{service}/user/{id}'.format(**favorite_user))
+        return
+    except:
+        print('Error getting favorite users. Session might have expired, login to kemono.party then get cookies.txt')
+        return
+
+def get_favorite_posts():
+    try:
+        api_favorites = 'https://kemono.party/api/favorites?type=post'
+        api_responce = requests.get(api_favorites, cookies=cookie_jar)
+        api_responce.raise_for_status()
+        data = json.loads(api_responce.text)
+        if not data:
+            print('You have no favorite posts.')
+            return
+        for favorite_post in data:
+            """
+            favorite_post       # dict, same as post
+                ['faved_seq']   # int
+            """
+            extract_link('https://kemono.party/{service}/user/{user}/post/{id}'.format(**favorite_post))
+        return
+    except:
+        print('Error getting favorite posts. Session might have expired, login to kemono.party then get cookies.txt')
+        return
+
 def main():
+    
+    if args['favorite_users']:
+        get_favorite_users()
+        
+    if args['favorite_posts']:
+        get_favorite_posts()
+        
     if args['links']:
         links = args['links'].split(",")
         for link in links:
